@@ -13,38 +13,38 @@
   "license": "MIT",
   "dependencies": {
     "@emotion/cache": "^11.4.0",
-    "@emotion/react": "^11.4.0",
+    "@emotion/react": "^11.4.1",
     "@emotion/server": "^11.4.0",
     "@emotion/styled": "^11.3.0",
-    "@material-ui/core": "next",
-    "@material-ui/icons": "next",
     "@mdx-js/react": "^1.6.22",
+    "@mui/icons-material": "^5.0.3",
+    "@mui/material": "^5.0.3",
     "action-reducer": "^0.3.1",
     "blob-stream": "^0.1.3",
     "keycode-js": "^3.1.0",
-    "next": "^11.0.1",
-    "pdfkit": "^0.12.1",
+    "next": "^11.1.2",
+    "pdfkit": "^0.12.3",
     "promise-worker": "^2.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-image-crop": "^8.6.12",
-    "recoil": "0.3.1"
+    "react-image-crop": "^9.0.4",
+    "recoil": "0.4.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.6",
+    "@babel/core": "^7.15.8",
     "@babel/preset-react": "^7.14.5",
     "@emotion/babel-preset-css-prop": "^11.2.0",
     "@mdx-js/loader": "^1.6.22",
-    "@next/mdx": "^11.0.1",
+    "@next/mdx": "^11.1.2",
     "@types/blob-stream": "^0.1.30",
-    "@types/pdfkit": "^0.11.1",
-    "@types/react": "^17.0.13",
-    "@types/react-dom": "^17.0.8",
-    "@types/react-image-crop": "^8.1.2",
-    "@types/react-modal": "^3.12.0",
+    "@types/pdfkit": "^0.11.2",
+    "@types/react": "^17.0.27",
+    "@types/react-dom": "^17.0.9",
+    "@types/react-image-crop": "^8.1.3",
+    "@types/react-modal": "^3.12.1",
     "@types/recoil": "0.0.1",
     "null-loader": "^4.0.1",
     "raw-loader": "^4.0.2",
-    "typescript": "^4.3.5"
+    "typescript": "^4.4.3"
   }
 }
