@@ -1,4 +1,4 @@
-import { AppBar, Toolbar, Typography } from '@material-ui/core'
+import { AppBar, Toolbar, Typography } from '@mui/material'
 import React from 'react'
 
 const Header = () =>
